Moved duplicate tests from MatFunctionsTests to the LnTests class.
