+++ EPPlusTest/FormulaParsing/Excel/Functions/MathFunctionsTests.cs
@@ -663,16 +663,6 @@ namespace EPPlusTest.Excel.Functions
 		}
 
 		[TestMethod]
-		public void LnShouldReturnCorrectResult()
-		{
-			var func = new Ln();
-			var args = FunctionsHelper.CreateArgs(5);
-			var result = func.Execute(args, _parsingContext);
-			var roundedResult = Math.Round((double)result.Result, 5);
-			Assert.AreEqual(1.60944d, roundedResult);
-		}
-
-		[TestMethod]
 		public void SqrtPiShouldReturnCorrectResult()
 		{
 			var func = new SqrtPi();
@@ -1194,16 +1184,6 @@ namespace EPPlusTest.Excel.Functions
 		public void LargeWithInvalidArgumentReturnsPoundValue()
 		{
 			var func = new Large();
-			var parsingContext = ParsingContext.Create();
-			var args = FunctionsHelper.CreateArgs();
-			var result = func.Execute(args, parsingContext);
-			Assert.AreEqual(eErrorType.Value, ((ExcelErrorValue)result.Result).Type);
-		}
-
-		[TestMethod]
-		public void LnWithInvalidArgumentReturnsPoundValue()
-		{
-			var func = new Ln();
 			var parsingContext = ParsingContext.Create();
 			var args = FunctionsHelper.CreateArgs();
 			var result = func.Execute(args, parsingContext);
